requirements.txt
django==3.2.6
djangorestframework==3.12.4
psycopg2==2.9.1